requirements.txt
Flask==3.0.2
python-dotenv==0.21.0